+++ DetectiveLog/DetectiveLog/TempDetail/ViewModel/TempDetailView.swift
